% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driving_accuracy.R
\name{driving_accuracy}
\alias{driving_accuracy}
\title{Calculate driving accuracy}
\usage{
driving_accuracy(fairways_hit, fairways_total)
}
\arguments{
\item{fairways_hit}{A number of fairways hit}

\item{fairways_total}{A number of total fairways}
}
\value{
Driving accuracy as a percentage
}
\description{
Calculate driving accuracy
}
