% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gir_rate.R
\name{gir_rate}
\alias{gir_rate}
\title{Calculate greens in regulation rate}
\usage{
gir_rate(gir_hit, holes)
}
\arguments{
\item{gir_hit}{Number of greens hit}

\item{holes}{Number of holes played}
}
\value{
GIR percentage
}
\description{
Calculate greens in regulation rate
}
