% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handicap_index.R
\name{handicap_index}
\alias{handicap_index}
\title{Calculate simplified golf handicap index}
\usage{
handicap_index(scores)
}
\arguments{
\item{scores}{Number of golf scores}
}
\value{
Estimated handicap index
}
\description{
Calculate simplified golf handicap index
}
