% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putt_efficiency.R
\name{putt_efficiency}
\alias{putt_efficiency}
\title{Putting performance}
\usage{
putt_efficiency(putts)
}
\arguments{
\item{putts}{Number of putts}
}
\value{
Summary of putting stats
}
\description{
Putting performance
}
