% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_trend.R
\name{round_trend}
\alias{round_trend}
\title{Track score trend over rounds}
\usage{
round_trend(rounds, scores)
}
\arguments{
\item{rounds}{Number of rounds}

\item{scores}{Number of scores}
}
\value{
Data frame of trends
}
\description{
Track score trend over rounds
}
