% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_avg.R
\name{score_avg}
\alias{score_avg}
\title{Calculate the average golf score}
\usage{
score_avg(scores)
}
\arguments{
\item{scores}{A number of golf scores}
}
\value{
The average score
}
\description{
Calculate the average golf score
}
